Tests/reorganize: move ICal tests to own file
