+++ test/PHPMailer/PHPMailerTest.php
@@ -1895,131 +1895,6 @@ EOT;
         $this->Mail->smtpClose();
     }
 
-    /**
-     * Test ICal method
-     */
-    public function testICalMethod()
-    {
-        $this->Mail->Subject .= ': ICal method';
-        $this->Mail->Body = '<h3>ICal method test.</h3>';
-        $this->Mail->AltBody = 'ICal method test.';
-        $this->Mail->Ical = 'BEGIN:VCALENDAR'
-            . "\r\nVERSION:2.0"
-            . "\r\nPRODID:-//PHPMailer//PHPMailer Calendar Plugin 1.0//EN"
-            . "\r\nMETHOD:CANCEL"
-            . "\r\nCALSCALE:GREGORIAN"
-            . "\r\nX-MICROSOFT-CALSCALE:GREGORIAN"
-            . "\r\nBEGIN:VEVENT"
-            . "\r\nUID:201909250755-42825@test"
-            . "\r\nDTSTART;20190930T080000Z"
-            . "\r\nSEQUENCE:2"
-            . "\r\nTRANSP:OPAQUE"
-            . "\r\nSTATUS:CONFIRMED"
-            . "\r\nDTEND:20190930T084500Z"
-            . "\r\nLOCATION:[London] London Eye"
-            . "\r\nSUMMARY:Test ICal method"
-            . "\r\nATTENDEE;CN=Attendee, Test;ROLE=OPT-PARTICIPANT;PARTSTAT=NEEDS-ACTION;RSVP="
-            . "\r\n TRUE:MAILTO:attendee-test@example.com"
-            . "\r\nCLASS:PUBLIC"
-            . "\r\nDESCRIPTION:Some plain text"
-            . "\r\nORGANIZER;CN=\"Example, Test\":MAILTO:test@example.com"
-            . "\r\nDTSTAMP:20190925T075546Z"
-            . "\r\nCREATED:20190925T075709Z"
-            . "\r\nLAST-MODIFIED:20190925T075546Z"
-            . "\r\nEND:VEVENT"
-            . "\r\nEND:VCALENDAR";
-        $this->buildBody();
-        $this->Mail->preSend();
-        self::assertMatchesRegularExpression(
-            '/Content-Type: text\/calendar; method=CANCEL;/',
-            $this->Mail->getSentMIMEMessage(),
-            'Wrong ICal method in Content-Type header'
-        );
-    }
-
-    /**
-     * Test ICal missing method to use default (REQUEST)
-     */
-    public function testICalInvalidMethod()
-    {
-        $this->Mail->Subject .= ': ICal method';
-        $this->Mail->Body = '<h3>ICal method test.</h3>';
-        $this->Mail->AltBody = 'ICal method test.';
-        $this->Mail->Ical = 'BEGIN:VCALENDAR'
-            . "\r\nVERSION:2.0"
-            . "\r\nPRODID:-//PHPMailer//PHPMailer Calendar Plugin 1.0//EN"
-            . "\r\nMETHOD:INVALID"
-            . "\r\nCALSCALE:GREGORIAN"
-            . "\r\nX-MICROSOFT-CALSCALE:GREGORIAN"
-            . "\r\nBEGIN:VEVENT"
-            . "\r\nUID:201909250755-42825@test"
-            . "\r\nDTSTART;20190930T080000Z"
-            . "\r\nSEQUENCE:2"
-            . "\r\nTRANSP:OPAQUE"
-            . "\r\nSTATUS:CONFIRMED"
-            . "\r\nDTEND:20190930T084500Z"
-            . "\r\nLOCATION:[London] London Eye"
-            . "\r\nSUMMARY:Test ICal method"
-            . "\r\nATTENDEE;CN=Attendee, Test;ROLE=OPT-PARTICIPANT;PARTSTAT=NEEDS-ACTION;RSVP="
-            . "\r\n TRUE:MAILTO:attendee-test@example.com"
-            . "\r\nCLASS:PUBLIC"
-            . "\r\nDESCRIPTION:Some plain text"
-            . "\r\nORGANIZER;CN=\"Example, Test\":MAILTO:test@example.com"
-            . "\r\nDTSTAMP:20190925T075546Z"
-            . "\r\nCREATED:20190925T075709Z"
-            . "\r\nLAST-MODIFIED:20190925T075546Z"
-            . "\r\nEND:VEVENT"
-            . "\r\nEND:VCALENDAR";
-        $this->buildBody();
-        $this->Mail->preSend();
-        self::assertMatchesRegularExpression(
-            '/Content-Type: text\/calendar; method=REQUEST;/',
-            $this->Mail->getSentMIMEMessage(),
-            'Wrong ICal method in Content-Type header'
-        );
-    }
-
-    /**
-     * Test ICal invalid method to use default (REQUEST)
-     */
-    public function testICalDefaultMethod()
-    {
-        $this->Mail->Subject .= ': ICal method';
-        $this->Mail->Body = '<h3>ICal method test.</h3>';
-        $this->Mail->AltBody = 'ICal method test.';
-        $this->Mail->Ical = 'BEGIN:VCALENDAR'
-            . "\r\nVERSION:2.0"
-            . "\r\nPRODID:-//PHPMailer//PHPMailer Calendar Plugin 1.0//EN"
-            . "\r\nCALSCALE:GREGORIAN"
-            . "\r\nX-MICROSOFT-CALSCALE:GREGORIAN"
-            . "\r\nBEGIN:VEVENT"
-            . "\r\nUID:201909250755-42825@test"
-            . "\r\nDTSTART;20190930T080000Z"
-            . "\r\nSEQUENCE:2"
-            . "\r\nTRANSP:OPAQUE"
-            . "\r\nSTATUS:CONFIRMED"
-            . "\r\nDTEND:20190930T084500Z"
-            . "\r\nLOCATION:[London] London Eye"
-            . "\r\nSUMMARY:Test ICal method"
-            . "\r\nATTENDEE;CN=Attendee, Test;ROLE=OPT-PARTICIPANT;PARTSTAT=NEEDS-ACTION;RSVP="
-            . "\r\n TRUE:MAILTO:attendee-test@example.com"
-            . "\r\nCLASS:PUBLIC"
-            . "\r\nDESCRIPTION:Some plain text"
-            . "\r\nORGANIZER;CN=\"Example, Test\":MAILTO:test@example.com"
-            . "\r\nDTSTAMP:20190925T075546Z"
-            . "\r\nCREATED:20190925T075709Z"
-            . "\r\nLAST-MODIFIED:20190925T075546Z"
-            . "\r\nEND:VEVENT"
-            . "\r\nEND:VCALENDAR";
-        $this->buildBody();
-        $this->Mail->preSend();
-        self::assertMatchesRegularExpression(
-            '/Content-Type: text\/calendar; method=REQUEST;/',
-            $this->Mail->getSentMIMEMessage(),
-            'Wrong ICal method in Content-Type header'
-        );
-    }
-
     public function testGivenIdnAddress_addAddress_returns_true()
     {
         if (file_exists(\PHPMAILER_INCLUDE_DIR . '/test/fakefunctions.php') === false) {
